+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## CHANGELOG
 
-### Next
+### 2.1.0
 
 - Add `REGEX_PICTO` which matches codepoints with the **Extended_Pictographic** property
 - Add `REGEX_PICTO_NO_EMOJI` which matches codepoints with the **Extended_Pictographic** property, but no **Emoji** property